requirements.txt
kafka-python==1.4.7
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.0
schedule==1.2.2
nest-asyncio==1.6.0
aiohttp==3.9.5
websockets==12.0
aiokafka==0.11.0
uvicorn==0.28.0
fastapi==0.110.1
SQLAlchemy==2.0.29
redis==5.0.7
aiomysql==0.2.0
cryptography==42.0.5
aioredis==2.0.1